+++ packages/web3-eth-contract/src/contract.ts
@@ -15,7 +15,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { Web3Context, Web3EventEmitter, Web3PromiEvent } from 'web3-core';
+import { Web3Context, Web3EventEmitter, Web3PromiEvent, Web3ConfigEvent } from 'web3-core';
 import { ContractExecutionError, SubscriptionError, Web3ContractError } from 'web3-errors';
 import {
 	createAccessList,
@@ -57,7 +57,6 @@ import {
 	HexString,
 	LogsInput,
 	Mutable,
-	Common,
 } from 'web3-types';
 import {
 	DataFormat,
@@ -213,69 +212,9 @@ export class Contract<Abi extends ContractAbi>
 	public readonly options: ContractOptions;
 
 	/**
-	 * Can be used to set {@link Contract.defaultAccount} for all contracts.
+	 * Set to true if you want contracts' defaults to sync with global defaults.
 	 */
-	public static defaultAccount?: HexString;
-
-	/**
-	 * Can be used to set {@link Contract.defaultBlock} for all contracts.
-	 */
-	public static defaultBlock?: BlockNumberOrTag;
-
-	/**
-	 * Can be used to set {@link Contract.defaultHardfork} for all contracts.
-	 */
-	public static defaultHardfork?: string;
-
-	/**
-	 * Can be used to set {@link Contract.defaultCommon} for all contracts.
-	 */
-	public static defaultCommon?: Common;
-
-	/**
-	 * Can be used to set {@link Contract.transactionSendTimeout} for all contracts.
-	 */
-	public static transactionSendTimeout?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionBlockTimeout} for all contracts.
-	 */
-	public static transactionBlockTimeout?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionConfirmationBlocks} for all contracts.
-	 */
-	public static transactionConfirmationBlocks?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionPollingInterval} for all contracts.
-	 */
-	public static transactionPollingInterval?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionPollingTimeout} for all contracts.
-	 */
-	public static transactionPollingTimeout?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionReceiptPollingInterval} for all contracts.
-	 */
-	public static transactionReceiptPollingInterval?: number;
-
-	/**
-	 * Can be used to set {@link Contract.transactionConfirmationPollingInterval} for all contracts.
-	 */
-	public static transactionConfirmationPollingInterval?: number;
-
-	/**
-	 * Can be used to set {@link Contract.blockHeaderTimeout} for all contracts.
-	 */
-	public static blockHeaderTimeout?: number;
-
-	/**
-	 * Can be used to set {@link Contract.handleRevert} for all contracts.
-	 */
-	public static handleRevert?: boolean;
+	public syncWithContext = false;
 
 	private _errorsInterface!: AbiErrorFragment[];
 	private _jsonInterface!: ContractAbiWithSignature;
@@ -292,6 +231,7 @@ export class Contract<Abi extends ContractAbi>
 	private _methods!: ContractMethodsInterface<Abi>;
 	private _events!: ContractEventsInterface<Abi>;
 
+	private context?: Web3Context;
 	/**
 	 * Creates a new contract instance with all its methods and events defined in its {@doclink glossary/json_interface | json interface} object.
 	 *
@@ -420,6 +360,8 @@ export class Contract<Abi extends ContractAbi>
 			data: options?.data,
 		};
 
+		this.syncWithContext = (options as ContractInitOptions)?.syncWithContext ?? false;
+
 		Object.defineProperty(this.options, 'address', {
 			set: (value: Address) => this._parseAndSetAddress(value, returnDataFormat),
 			get: () => this._address,
@@ -429,131 +371,6 @@ export class Contract<Abi extends ContractAbi>
 			set: (value: ContractAbi) => this._parseAndSetJsonInterface(value, returnDataFormat),
 			get: () => this._jsonInterface,
 		});
-	}
-
-	public get defaultAccount() {
-		return (this.constructor as typeof Contract).defaultAccount ?? super.defaultAccount;
-	}
-
-	public set defaultAccount(value: Address | undefined) {
-		super.defaultAccount = value;
-	}
-
-	public get defaultBlock() {
-		return (this.constructor as typeof Contract).defaultBlock ?? super.defaultBlock;
-	}
-
-	public set defaultBlock(value: BlockNumberOrTag) {
-		super.defaultBlock = value;
-	}
-
-	public get defaultHardfork() {
-		return (this.constructor as typeof Contract).defaultHardfork ?? super.defaultHardfork;
-	}
-
-	public set defaultHardfork(value: string) {
-		super.defaultHardfork = value;
-	}
-
-	public get defaultCommon(): Common | undefined {
-		return (this.constructor as typeof Contract).defaultCommon ?? super.defaultCommon;
-	}
-
-	public set defaultCommon(value: Common | undefined) {
-		super.defaultCommon = value;
-	}
-
-	public get transactionSendTimeout() {
-		return (
-			(this.constructor as typeof Contract).transactionSendTimeout ??
-			super.transactionSendTimeout
-		);
-	}
-
-	public set transactionSendTimeout(value: number) {
-		super.transactionSendTimeout = value;
-	}
-
-	public get transactionBlockTimeout() {
-		return (
-			(this.constructor as typeof Contract).transactionBlockTimeout ??
-			super.transactionBlockTimeout
-		);
-	}
-
-	public set transactionBlockTimeout(value: number) {
-		super.transactionBlockTimeout = value;
-	}
-
-	public get transactionConfirmationBlocks() {
-		return (
-			(this.constructor as typeof Contract).transactionConfirmationBlocks ??
-			super.transactionConfirmationBlocks
-		);
-	}
-
-	public set transactionConfirmationBlocks(value: number) {
-		super.transactionConfirmationBlocks = value;
-	}
-
-	public get transactionPollingInterval() {
-		return (
-			(this.constructor as typeof Contract).transactionPollingInterval ??
-			super.transactionPollingInterval
-		);
-	}
-
-	public set transactionPollingInterval(value: number) {
-		super.transactionPollingInterval = value;
-	}
-
-	public get transactionPollingTimeout() {
-		return (
-			(this.constructor as typeof Contract).transactionPollingTimeout ??
-			super.transactionPollingTimeout
-		);
-	}
-
-	public set transactionPollingTimeout(value: number) {
-		super.transactionPollingTimeout = value;
-	}
-
-	public get transactionReceiptPollingInterval() {
-		return (
-			(this.constructor as typeof Contract).transactionReceiptPollingInterval ??
-			super.transactionReceiptPollingInterval
-		);
-	}
-
-	public set transactionReceiptPollingInterval(value: number | undefined) {
-		super.transactionReceiptPollingInterval = value;
-	}
-
-	public get transactionConfirmationPollingInterval() {
-		return (
-			(this.constructor as typeof Contract).transactionConfirmationPollingInterval ??
-			super.transactionConfirmationPollingInterval
-		);
-	}
-
-	public set transactionConfirmationPollingInterval(value: number | undefined) {
-		super.transactionConfirmationPollingInterval = value;
-	}
-
-	public get blockHeaderTimeout() {
-		return (this.constructor as typeof Contract).blockHeaderTimeout ?? super.blockHeaderTimeout;
-	}
-
-	public set blockHeaderTimeout(value: number) {
-		super.blockHeaderTimeout = value;
-	}
-
-	public get handleRevert() {
-		return (this.constructor as typeof Contract).handleRevert ?? super.handleRevert;
-	}
-
-	public set handleRevert(value: boolean) {
-		super.handleRevert = value;
 	}
 
 	/**
@@ -633,8 +450,10 @@ export class Contract<Abi extends ContractAbi>
 	 * ```
 	 */
 	public clone() {
+		let newContract: Contract<any>;
+
 		if (this.options.address) {
-			return new Contract<Abi>(
+			newContract = new Contract<Abi>(
 				[...this._jsonInterface, ...this._errorsInterface] as unknown as Abi,
 				this.options.address,
 				{
@@ -644,23 +463,28 @@ export class Contract<Abi extends ContractAbi>
 					from: this.options.from,
 					data: this.options.data,
 					provider: this.currentProvider,
+					syncWithContext: this.syncWithContext,
+				},
+				this.getContextObject(),
+			);
+		} else {
+			newContract = new Contract<Abi>(
+				[...this._jsonInterface, ...this._errorsInterface] as unknown as Abi,
+				{
+					gas: this.options.gas,
+					gasPrice: this.options.gasPrice,
+					gasLimit: this.options.gasLimit,
+					from: this.options.from,
+					data: this.options.data,
+					provider: this.currentProvider,
+					syncWithContext: this.syncWithContext,
 				},
 				this.getContextObject(),
 			);
 		}
+		if (this.context) newContract.subscribeToContextEvents(this.context);
 
-		return new Contract<Abi>(
-			[...this._jsonInterface, ...this._errorsInterface] as unknown as Abi,
-			{
-				gas: this.options.gas,
-				gasPrice: this.options.gasPrice,
-				gasLimit: this.options.gasLimit,
-				from: this.options.from,
-				data: this.options.data,
-				provider: this.currentProvider,
-			},
-			this.getContextObject(),
-		);
+		return newContract;
 	}
 
 	/**
@@ -772,6 +596,7 @@ export class Contract<Abi extends ContractAbi>
 				// modifiedOptions.to = '0x0000000000000000000000000000000000000000';
 				delete modifiedOptions.to;
 
+				// eslint-disable-next-line @typescript-eslint/no-unsafe-return
 				return this._contractMethodDeploySend(
 					abi as AbiFunctionFragment,
 					args as unknown[],
@@ -1329,5 +1154,17 @@ export class Contract<Abi extends ContractAbi>
 
 			return sub;
 		};
+	}
+
+	protected subscribeToContextEvents<T extends Web3Context>(context: T): void {
+		// eslint-disable-next-line @typescript-eslint/no-this-alias
+		const contractThis = this;
+		this.context = context;
+
+		if (contractThis.syncWithContext) {
+			context.on(Web3ConfigEvent.CONFIG_CHANGE, event => {
+				contractThis.setConfig({ [event.name]: event.newValue });
+			});
+		}
 	}
 }