+++ packages/web3-eth-contract/CHANGELOG.md
@@ -216,6 +216,10 @@ const transactionHash = receipt.transactionHash;
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix contract defaults (#5756)
+
 ### Changed
 
 -   Update imports statements for objects that was moved between web3 packages (#5771)